Split out message parsing
