+++ src/features/jobs-moderation/parse-content.ts
@@ -1,0 +1,36 @@
+import { Post } from "./job-mod-helpers";
+import { parseTags } from "./formatting";
+
+export const parseContent = (inputString: string): Post[] => {
+  const lines = inputString.trim().split("\n");
+  const posts = lines.reduce<Post[]>((acc, line) => {
+    if (line === "") return acc;
+    if (line.includes("|") || line.includes("[")) {
+      // This line contains tags, so we're starting a new post
+      acc.push({
+        tags: parseTags(line),
+        description: "",
+        // contact: "",
+      });
+    } else {
+      // This line belongs to the current post
+      const currentPost = acc[acc.length - 1];
+      // If there are no tags, we should still parse correctly
+      if (!currentPost) {
+        acc.push({ tags: [], description: line });
+        // } else if (currentPost.contact === "") {
+        // currentPost.contact = "";
+        // } else {
+        // currentPost.contact += "\n" + line;
+      } else if (currentPost.description === "") {
+        currentPost.description += line;
+      } else {
+        currentPost.description += "\n" + line;
+      }
+    }
+    return acc;
+  }, []);
+  return posts.map((post) => {
+    return { ...post };
+  });
+};
